+++ ignite/pkg/chaincmd/runner/chain.go
@@ -57,6 +57,19 @@ func (r Runner) InPlace(ctx context.Context, newChainID, newOperatorAddress stri
 	)
 }
 
+// Initialize config directories & files for a multi-validator testnet locally.
+func (r Runner) MultiNode(ctx context.Context, options ...chaincmd.MultiNodeOption) error {
+	runOptions := runOptions{
+		stdout: os.Stdout,
+		stderr: os.Stderr,
+	}
+	return r.run(
+		ctx,
+		runOptions,
+		r.chainCmd.TestnetMultiNodeCommand(options...),
+	)
+}
+
 // Gentx generates a genesis tx carrying a self delegation.
 func (r Runner) Gentx(
 	ctx context.Context,